test(webapp): split runsReplicationService into 6 files (per-test-pg tests ~29s; keep each shard <2m)
